refactor by moving specific code into its own file
